ARM: Add i9300 and i9305 defconfig to create a zImage-dtb
Currently the AOSP or LineageOS build configuration lack
a way to append a dtb to the kernel image[1].

This adds the necessary configuration to do that, which
depends on the following commit:
  93c1eb8ea30b ANDROID: ARM: convert build of appended dtb zImage to list of dtbs
  c1cb2ea535c5 ANDROID: ARM: add config option to build zImage/dtb combo

The defconfig are based on midas_defconfig, which can be found in
the following commit by Joonas Kylmälä:
  f236e10fabb6 Make AOSP defconfig for Midas devices
which is available in the following repository:
  git://git.putti.eu/aosp/kernel_i9305.git

The differences with that configuration are minimal.
For the i9300 we have:
  $ diff -u midas_defconfig replicant_i9300_defconfig
  +CONFIG_BUILD_ARM_APPENDED_DTB_IMAGE=y
  +CONFIG_BUILD_ARM_APPENDED_DTB_IMAGE_NAMES="exynos4412-i9300"
and for i9305:
  $ diff -u midas_defconfig replicant_i9305_defconfig
  +CONFIG_BUILD_ARM_APPENDED_DTB_IMAGE=y
  +CONFIG_BUILD_ARM_APPENDED_DTB_IMAGE_NAMES="exynos4412-i9305"

Ideally this should be instead handled outside of the kernel, for
instance in a bootloader or in the build system assuming there is
in a way to do it that can easily be maintained and rebased.

The configurations use CONFIG_CMDLINE because adding the
command line options in the boot.img[2] has no effect as the
proprietary samsung bootloader[3] ignores and but still sets
its own command line arguments:
- We can't find any trace of the cmdline we added in the
  boot.img in s-boot's ATAG_CMDLINE. During the boot,
  that CMDLINE is printed on the serial console if you
  hold the volume-down button:
    ATAG_CMDLINE: 7f 54410009 'console=ram loglevel=4 [...]'
- We can't find any trace of the cmdline added in boot.img
  in the kernel boot log either, while we can find the
  cmdline added by s-boot.

References:
-----------
[1]in vendor/lineage/build/soong/Android.bp and
   vendor/lineage/build/tasks/kernel.mk
[2]This was validated by passing the command line opions to
   mkbootimg through --cmdline while manually generating a
   boot.img
   The mkbootimg used came from android-tools 9.0.0_r35-1.0
   which came from Parabola/
[3]This was validated on a Galaxy SIII with this s-boot version:
   Samsung S-Boot 4.0-2713958 for GT-I9300 (Jul 16 2014 - 18:09:55)

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>