Use serial console for now
The use of CONFIG_CMDLINE_FORCE should be revisited, and using
the Android bootimage or some other means to set the cmdline
could be used if the bootloader that is used doesn't stand in
the way.

This earlyprintk option was removed as it is not necessary
anymore. The earlyprintk option is typicaly used when the
serial console driver is not working properly or when the
kernel is crashing before having the opportunity to enable
the serial console driver.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>