HACK: power_supply: make prop type report real values
This is required for Android healthd to distinguish
between the battery and the charger power supplies.
Note that the smdk4412 kernel returns "Battery" in the
case of max77693 charger, and the healthd still able
to detect the charging somehow.

TODO: investigate how this works on stock 3.0.101 kernel,
and remove this hack.